Continued refactoring Pipe and added builder
Continued to move logic out of the Pipe class
and added a builder class that will be used instead of
inheriting from the Pipe class. Essentially,
functions will use the new PipeBuilder to
create and return custom instances of the Pipe class.

The PipeBuilder is implemented as a friend of the Pipe
class so that it can expose more methods during construction
but still hide things that shouldn't be touched.